# backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/preprocess/{uuid}:
    post:
      tags:
      - preprocess-controller
      operationId: preprocess
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskAckResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/preprocess/{uuid}/abort:
    post:
      tags:
      - preprocess-controller
      operationId: abortTask
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskAckResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/preprocess/{uuid}/progress:
    get:
      tags:
      - preprocess-controller
      operationId: getProgress
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskProgress"
  /api/preprocess/images/{id}/file:
    get:
      tags:
      - preprocess-controller
      operationId: getImageFileById
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
  /api/preprocess/images/summaries:
    get:
      tags:
      - preprocess-controller
      operationId: getImageSummariesByUuid
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageImageSummaryEntity"
    delete:
      tags:
      - preprocess-controller
      operationId: deleteImageSummariesByIds
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/preprocess/audios/{id}/file:
    get:
      tags:
      - preprocess-controller
      operationId: getAudioFileById
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
  /api/preprocess/audios/summaries:
    get:
      tags:
      - preprocess-controller
      operationId: getAudioSummariesByUuid
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageAudioSummary"
    delete:
      tags:
      - preprocess-controller
      operationId: deleteAudioSummariesByIds
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/preprocess/images/summaries/{id}:
    delete:
      tags:
      - preprocess-controller
      operationId: deleteImageSummaryById
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/preprocess/images/summaries/all:
    delete:
      tags:
      - preprocess-controller
      operationId: deleteAllImageSummariesByUuid
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/preprocess/audios/summaries/{id}:
    delete:
      tags:
      - preprocess-controller
      operationId: deleteAudioSummaryById
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/preprocess/audios/summaries/{id}/text:
    delete:
      tags:
      - preprocess-controller
      operationId: clearAudioSummaryTextById
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/preprocess/audios/summaries/text:
    delete:
      tags:
      - preprocess-controller
      operationId: clearAudioSummaryTextsByIds
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/preprocess/audios/summaries/all:
    delete:
      tags:
      - preprocess-controller
      operationId: deleteAllAudioSummariesByUuid
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
components:
  schemas:
    ErrorResponse:
      type: object
      description: Unified error body. The code mirrors the HTTP status.
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    ApiResponsePageAudioSummary:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageAudioSummary"
    ApiResponsePageImageSummaryEntity:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageImageSummaryEntity"
...
    ApiResponseTaskAckResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/TaskAckResponse"
    ApiResponseTaskProgress:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/TaskProgress"
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    TaskAckResponse:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
          nullable: true
          enum:
          - INITIAL_STATE
          - RUNNING
          - PAUSED
          - SUCCESS
          - FAILED
        message:
          type: string
...
    TaskProgress:
      type: object
      properties:
        status:
          type: string
          enum:
          - IDLING
          - RUNNING
          - PAUSED
          - COMPLETED
        totalTasks:
          type: integer
          format: int32
        remainingTasks:
          type: integer
          format: int32
        completedTasks:
          type: integer
          format: int32
        progressPercentage:
          type: number
          format: double
    ImageSummaryEntity:
      type: object
      properties:
        id:
          type: string
        imageHash:
          type: string
        filePath:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
    PageImageSummaryEntity:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ImageSummaryEntity"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
    AudioSummary:
      type: object
      properties:
        id:
          type: string
        fileHash:
          type: string
        filePath:
          type: string
        transcript:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
    PageAudioSummary:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/AudioSummary"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...